Extracts ListContent from ListCollectibles
